make header-serach-bar-spec tests- solve complexity header
